fix lint.
git-svn-id: file:///svn/unbound/trunk@2545 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ winrc/anchor-update.c
@@ -85,7 +85,7 @@ result2answer(struct ub_result* result)
 {
 	ldns_pkt* p = NULL;
 	ldns_rr_list* a;
-	if(ldns_wire2pkt(&p, result->answer_packet, result->answer_len) 
+	if(ldns_wire2pkt(&p, result->answer_packet, (size_t)result->answer_len) 
 		!= LDNS_STATUS_OK) 
 		return NULL;
 	a = ldns_pkt_answer(p);
